+++ deltafi-ui/src/components/errors/SummaryByFlowPanel.vue
@@ -24,7 +24,7 @@
         <span class="fas fa-bars" />
       </Button>
       <Menu ref="menu" :model="menuItems" :popup="true" />
-      <Paginator v-if="errorsFlow.length > 0" :rows="10" template="CurrentPageReport FirstPageLink PrevPageLink PageLinks NextPageLink LastPageLink RowsPerPageDropdown" current-page-report-template="{first} - {last} of {totalRecords}" :total-records="totalErrorsFlow" :rows-per-page-options="[10, 20, 50, 100, 1000]" class="p-panel-header" style="float: left" @page="onPage($event)"></Paginator>
+      <Paginator v-if="errorsFlow.length > 0" :rows="perPage" template="CurrentPageReport FirstPageLink PrevPageLink PageLinks NextPageLink LastPageLink RowsPerPageDropdown" current-page-report-template="{first} - {last} of {totalRecords}" :total-records="totalErrorsFlow" :rows-per-page-options="[10, 20, 50, 100, 1000]" class="p-panel-header" style="float: left" @page="onPage($event)"></Paginator>
     </template>
     <DataTable id="errorsSummaryTable" v-model:selection="selectedErrors" responsive-layout="scroll" selection-mode="multiple" data-key="flow" class="p-datatable-gridlines p-datatable-sm" striped-rows :meta-key-selection="false" :value="errorsFlow" :loading="loading" :rows="perPage" :lazy="true" :total-records="totalErrorsFlow" :row-hover="true" @row-contextmenu="onRowContextMenu" @sort="onSort($event)">
       <template #empty>No results to display.</template>
@@ -51,14 +51,15 @@ import MetadataDialog from "@/components/MetadataDialog.vue";
 import useNotifications from "@/composables/useNotifications";
 import useUtilFunctions from "@/composables/useUtilFunctions";
 import AcknowledgeErrorsDialog from "@/components/AcknowledgeErrorsDialog.vue";
-import { ref, onMounted, defineExpose, defineProps, watch, computed, defineEmits } from "vue";
+import { ref, onMounted, defineExpose, defineProps, watch, computed, defineEmits, nextTick } from "vue";
+import { useStorage, StorageSerializers } from "@vueuse/core";
 
 const loading = ref(true);
 const menu = ref();
 const errorsFlow = ref([]);
 const totalErrorsFlow = ref(0);
 const offset = ref(0);
-const perPage = ref(10);
+const perPage = ref();
 const sortField = ref("modified");
 const metadataDialog = ref();
 const sortDirection = ref("DESC");
@@ -124,6 +125,7 @@ const menuItems = ref([
 ]);
 
 onMounted(() => {
+  getPersistedParams();
   fetchErrorsFlow();
 });
 
@@ -135,6 +137,7 @@ const onRefresh = () => {
 const { data: response, fetchByFlow: getErrorsByFlow } = useErrorsSummary();
 
 const fetchErrorsFlow = async () => {
+  getPersistedParams();
   let ingressFlowName = props.ingressFlowName != null ? props.ingressFlowName.name : null;
   let showAcknowled = props.awknowledged ? null : false;
   loading.value = true;
@@ -194,10 +197,13 @@ const onSort = (event) => {
   fetchErrorsFlow();
 };
 
-const onPage = (event) => {
+const onPage = async (event) => {
   offset.value = event.first;
   perPage.value = event.rows;
+  await nextTick();
+  setPersistedParams();
   fetchErrorsFlow();
+  emit("refreshErrors");
 };
 
 watch(
@@ -214,6 +220,18 @@ watch(
     fetchErrorsFlow();
   }
 );
+
+const getPersistedParams = async () => {
+  let state = useStorage("errors-page-session-storage", {}, sessionStorage, { serializer: StorageSerializers.object });
+  perPage.value = state.value.perPage || 10;
+};
+
+const setPersistedParams = () => {
+  let state = useStorage("errors-page-session-storage", {}, sessionStorage, { serializer: StorageSerializers.object });
+  state.value = {
+    perPage: perPage.value,
+  };
+};
 </script>
 
 <style lang="scss"></style>